♻️ ZXJVC7 pr: decompose protected integrate handoff
