+++ .agentplane/tasks/202605290241-ZXJVC7/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290241-ZXJVC7`
+Title: PR integrate command decomposition
+Canonical task record: `.agentplane/tasks/202605290241-ZXJVC7/README.md`
+
+## Summary
+
+PR integrate command decomposition
+
+Decompose packages/agentplane/src/commands/pr/integrate/cmd.ts by extracting protected-base GitHub PR handoff/merge routing while preserving integrate behavior and diagnostics.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/pr/integrate/cmd.ts by extracting protected-base GitHub PR handoff/merge routing while preserving integrate behavior and diagnostics.
+- Out of scope: unrelated refactors not required for "PR integrate command decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T02:42:28.306Z
+- Branch: task/202605290241-ZXJVC7/pr-integrate-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
