+++ .agentplane/tasks/202605290241-ZXJVC7/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 ZXJVC7 task: PR integrate command decomposition [202605290241-ZXJVC7]
