+++ .agentplane/tasks/202605290241-ZXJVC7/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T02:42:28.306Z
+
+## Task
+
+- Task: `202605290241-ZXJVC7`
+- Title: PR integrate command decomposition
+- Status: DOING
+- Branch: `task/202605290241-ZXJVC7/pr-integrate-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290241-ZXJVC7/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T02:42:28.306Z
+- Branch: task/202605290241-ZXJVC7/pr-integrate-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
